+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Tasks are now queued
 - Project structure improved
 - Live Reload setting only shows on iOS/Android projects
+- Fix for React and Vue projects to open browser when run on web
 
 ### Version 1.20.0